// src/app/generated/tslibs.ts
// Do not edit. Generated by misc/build-tslibs.js
  export const dom = {
  filename: "lib.dom.d.ts?raw",
  version: "4hG/R46RQSoRnGRlLKMnhD9gaX4"
};
export const webworker = {
  filename: "lib.webworker.d.ts?raw",
  version: "JvTccpIu9of2OjFw8xrHYjbBrVU"
};
export const scripter = {
  filename: "scripter-env.d.ts?raw",
  version: "OTv3981fJru4/uRsfKhUXweO9a0"
};
export const figma = {
  filename: "figma.d.ts?raw",
  version: "SPhe1CAPkagGPjiJ0QbGoQW5RvI"
};